pyproject.toml
[project]
name = "stake-allocation-simulation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
    "web3>=6.14.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "httpx>=0.24",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
markers = [
    "integration: marks tests that hit live network and may be slow",
]
